test(e2e): use a different approach
Instead of having e2e test per rule, (which did not make much sense as we
already have unit tests), we now simulate a SvelteKit app structure,
with a `tailwind.config.ts` file and other "real-world" assets.
We then still work with snapshots, as it's the easiest way to see the
issues in case a rule has been changed and there's an error due to that
change. It also allows to test a mix of rules in the same test.